+++ src/view/elements/private/HsmStateTextItem.hpp
@@ -3,6 +3,7 @@
 
 #include <QGraphicsTextItem>
 #include <QObject>
+#include <QPointer>
 
 namespace view {
 
@@ -10,9 +11,11 @@ class HsmStateTextItem : public QGraphicsTextItem {
     Q_OBJECT
 
 public:
-    HsmStateTextItem(QGraphicsItem* parent = nullptr);
+    // Logical parent is used for selection purposes when TextItem get's the focus
+    HsmStateTextItem(QGraphicsItem* parent, QGraphicsItem* logicalParent);
 
     void makeMovable(const bool enable);
+    void link(const QPointer<HsmStateTextItem>& other);
 
 signals:
     void editingFinished();
@@ -24,8 +27,11 @@ protected:
     void keyPressEvent(QKeyEvent* event) override;
     QVariant itemChange(GraphicsItemChange change, const QVariant& value) override;
 
+    void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget = nullptr) override;
+
 private:
     QString mOriginalText;
+    QGraphicsItem* mLogicalParent = nullptr;
 };
 
 }  // namespace view