+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to project will be documented in this file.
 
+## [0.2.0] - 2025-11-30
+### New
+- Rendering of transition callback and condition
+
+### Fixed
+- Improved label positioning for transitions
+- Fixed HsmGraphicsView not giving some key presses to child elements (couldn't type spaces in labels)
+
 ## [0.0.1] - 2025-11-29
 ### New
 - Start of the project versioning